Fix sign-compare warnings

+++ src/pctcp.c
@@ -3090,7 +3090,7 @@ void W32_CALL sock_flush (sock_type *s)
     _tcp_Socket *tcp = &s->tcp;
 
     tcp->sockmode &= ~SOCK_MODE_LOCAL;
-    if (tcp->tx_datalen > tcp->send_una)
+    if (tcp->tx_datalen > (unsigned long)tcp->send_una)
     {
       tcp->flags |= tcp_FlagPUSH;
       if (s->tcp.send_una == 0)  /* !! S. Lawson - only if data not moving */
